+++ Changelog.md
@@ -7,7 +7,7 @@
 - Remove the Stats panel and merge it into the Workspace panel
 - Add documentation into the About panel
 - Add random triangle color mesh render mode
-- Persist panel layout between sessions. I also added a button to the File menu to reset the UI layout.
+- Persist panel layout between sessions. I also added a button to reset the UI layout in the Workspace panel.
 
 ## v0.1.0 &mdash; Feb 12, 2025